refactor: implementing review comments
